refactor(mm): merge arch-dependent virtualmem modules
