+++ src/arch/aarch64/mm/mod.rs
@@ -1,6 +1,5 @@
 pub mod paging;
 pub mod physicalmem;
-pub mod virtualmem;
 
 pub use self::physicalmem::init_page_tables;
 
@@ -9,5 +8,5 @@ pub fn init() {
 		paging::init();
 	}
 	physicalmem::init();
-	virtualmem::init();
+	crate::mm::virtualmem::init();
 }
