+++ src/arch/aarch64/mm/paging.rs
@@ -8,7 +8,6 @@ use align_address::Align;
 use memory_addresses::{PhysAddr, VirtAddr};
 
 use crate::arch::aarch64::kernel::{get_base_address, get_image_size, get_ram_address, processor};
-use crate::arch::aarch64::mm::virtualmem;
 use crate::env::is_uhyve;
 use crate::mm::physicalmem;
 use crate::{KERNEL_STACK_SIZE, mm, scheduler};
