+++ src/arch/riscv64/mm/mod.rs
@@ -1,10 +1,9 @@
 pub mod paging;
 pub mod physicalmem;
-pub mod virtualmem;
 
 pub use self::paging::init_page_tables;
 
 pub fn init() {
 	physicalmem::init();
-	virtualmem::init();
+	crate::mm::virtualmem::init();
 }
